requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.2
async-timeout==4.0.1
attrs==21.2.0
certifi==2018.11.29
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
Django==3.2.11
djangorestframework==3.12.2
django-classy-tags==0.8.0
django-extensions==3.1.5
drf-yasg==1.20.0
marshmallow==3.14.1
marshmallow-dataclass==8.5.3
marshmallow-enum==1.5.1
PyYAML==6.0
et-xmlfile==1.1.0
fastapi==0.88.0
Flask==2.0.2
frozenlist==1.2.0
gevent==21.8.0
greenlet==1.1.2
h11==0.14.0
idna==2.8
itsdangerous==2.0.1
jdcal==1.4.1
Jinja2==3.0.3
MarkupSafe==2.0.1
multidict==5.2.0
numpy==1.20.3
opencv-python
openpyxl==2.6.2
pycparser==2.21
pydantic==1.10.2
pyTelegramBotAPI==4.8.0
python-dateutil==2.8.0
python-json-logger==2.0.2
pytz==2018.9
requests==2.21.0
six==1.12.0
sniffio==1.3.0
starlette==0.22.0
timestamp==0.0.1
typing-extensions==4.4.0
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==1.0.23
urllib3==1.24.1
uvicorn==0.20.0
webencodings==0.5.1
Werkzeug==2.0.2
xlwt==1.3.0
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
